Add send_data function to proto service
Allows to publish a message on Astarte with send_data function,
at this moment only Double and Int32 are supported, but this is
only example for trying to publish data on Astarte through the MessageHub.
Inject node-id and node-name to all node request in order to check which node
is trying to do something.
Rename HubMessage to NodeMessage.